Organize sidebar extension sources into folders
