+++ Packages/CmuxExtensionKit/Sources/CmuxExtensionKit/HostCompatibility/CmuxExtensionSidebarProvider.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public protocol CmuxExtensionSidebarProvider: Sendable {
+    var descriptor: CmuxExtensionSidebarProviderDescriptor { get }
+}
