+++ Packages/CmuxExtensionKit/Sources/CmuxExtensionKit/HostCompatibility/CmuxExtensionSidebarProviderID.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+// Compatibility name used only by the existing cmux app while the old prototype
+// sidebar is removed in smaller follow-up steps.
+public enum CmuxExtensionSidebarProviderID {
+    public static let defaultWorkspaces = "cmux.sidebar.default"
+}
