+++ Packages/CmuxExtensionKit/Sources/CmuxExtensionKit/Manifest/CMUXExtensionKind.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public enum CMUXExtensionKind: String, Codable, CaseIterable, Equatable, Sendable {
+    case sidebar
+}
