+++ Packages/CmuxExtensionKit/Sources/CmuxExtensionKit/Manifest/CMUXExtensionManifest.swift
@@ -1,9 +1,5 @@
 import Foundation
 
-public enum CMUXExtensionKind: String, Codable, CaseIterable, Equatable, Sendable {
-    case sidebar
-}
-
 public struct CMUXExtensionManifest: Codable, Equatable, Identifiable, Sendable {
     public var id: String
     public var displayName: String
@@ -24,12 +20,4 @@ public struct CMUXExtensionManifest: Codable, Equatable, Identifiable, Sendable 
         self.minimumAPIVersion = minimumAPIVersion
         self.requestedScopes = requestedScopes
     }
-}
-
-public enum CMUXExtensionScope: String, Codable, CaseIterable, Equatable, Sendable {
-    case workspaceMetadata
-    case workspacePaths
-    case notifications
-    case networkPorts
-    case pullRequests
 }
