+++ Packages/CmuxExtensionKit/Sources/CmuxExtensionKit/Manifest/CMUXExtensionScope.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+public enum CMUXExtensionScope: String, Codable, CaseIterable, Equatable, Sendable {
+    case workspaceMetadata
+    case workspacePaths
+    case notifications
+    case networkPorts
+    case pullRequests
+}
