+++ Packages/CmuxExtensionKit/Sources/CmuxExtensionKit/Sidebar/CMUXExtensionActionResult.swift
@@ -1,11 +1,5 @@
 import Foundation
 
-public enum CMUXSidebarAction: Codable, Equatable, Sendable {
-    case selectWorkspace(UUID)
-    case closeWorkspace(UUID)
-    case openURL(String)
-}
-
 public struct CMUXExtensionActionResult: Codable, Equatable, Sendable {
     public var accepted: Bool
     public var message: String?
