+++ Packages/CmuxExtensionKit/Sources/CmuxExtensionKit/Sidebar/CMUXSidebarAction.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+public enum CMUXSidebarAction: Codable, Equatable, Sendable {
+    case selectWorkspace(UUID)
+    case closeWorkspace(UUID)
+    case openURL(String)
+}
