+++ Packages/CmuxExtensionKit/Sources/CmuxExtensionKit/Validation/CMUXExtensionValidationError.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+public enum CMUXExtensionValidationError: Error, Equatable, Sendable {
+    case unsupportedKind(CMUXExtensionKind)
+    case unsupportedAPIVersion(requested: CMUXExtensionAPIVersion, supported: CMUXExtensionAPIVersion)
+    case emptyIdentifier
+    case emptyDisplayName
+}
