+++ Packages/CmuxExtensionKit/Sources/CmuxExtensionKit/Validation/CMUXExtensionValidator.swift
@@ -1,12 +1,5 @@
 import Foundation
 
-public enum CMUXExtensionValidationError: Error, Equatable, Sendable {
-    case unsupportedKind(CMUXExtensionKind)
-    case unsupportedAPIVersion(requested: CMUXExtensionAPIVersion, supported: CMUXExtensionAPIVersion)
-    case emptyIdentifier
-    case emptyDisplayName
-}
-
 public enum CMUXExtensionValidator {
     public static func validateSidebarManifest(
         _ manifest: CMUXExtensionManifest,
